// src/enums/admin.enum.ts
enum STATUS {
    ACTIVE = "ACTIVE",
    INACTIVE = "INACTIVE"
}

enum CHECKIN_STATUS {
    CHECKIN = "CHECKIN",
    CHECKOUT = "CHECKOUT"
}

enum ROLES {
    ADMIN = "ADMIN",
    SUB_ADMIN = "SUB_ADMIN",
    OPERATOR = "OPERATOR",
    CSR = "CSR",
    FINANCE = "FINANCE",
    RECEPTIONIST = "RECEPTIONIST"
}


export {
    STATUS,
    CHECKIN_STATUS,
    ROLES
}